requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.5
python-dotenv==1.0.1